processing/requirements.txt
pandas==2.0.3
boto3==1.28.25
pyarrow==12.0.0
numpy==1.24.3
scikit-learn==1.3.0
pydantic==2.1.1
python-dotenv==1.0.0